add authentication ui project #1
