+++ Avhrm.Core/Contracts/IAuthenticationService.cs
@@ -1,0 +1,10 @@
+﻿using System.ServiceModel;
+
+namespace Avhrm.Identity.Contracts;
+
+[ServiceContract]
+public interface IAuthenticationService
+{
+    [OperationContract]
+    Task<bool> Authenticate(string username, string password);
+}
