+++ Avhrm.Server/Startup.cs
@@ -1,6 +1,9 @@
 ﻿using Avhrm.Server.Intercepters;
 using Avhrm.Persistence;
 using Avhrm.Core;
+using Avhrm.Identity.Server;
+using Avhrm.Persistence.Repositories;
+using Avhrm.Identity.Server.Implementation;
 
 namespace Avhrm.Server;
 
@@ -15,6 +18,8 @@ public static class Startup
         services.AddCoreServices();
 
         services.AddPersistenceServices(configuration);
+
+        services.AddIdentityServerServices();
     }
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -36,6 +41,11 @@ public static class Startup
        
         app.UseAuthorization();
 
-        app.UsePersistenceMiddlewares();
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapGrpcService<VacationRequestRepository>();
+
+            endpoints.MapGrpcService<AuthenticationService>();
+        });
     }
 }
