+++ Avhrm.UI.Shared/SharedServices.cs
@@ -1,4 +1,6 @@
 ﻿using Avhrm.Core.Contracts;
+using Avhrm.Identity;
+using Avhrm.Identity.Contracts;
 using Avhrm.UI.Shared.Extensions;
 
 namespace Avhrm.UI.Shared;
@@ -8,5 +10,9 @@ public static class SharedServices
     public static void AddSharedServices(this IServiceCollection services)
     {
         services.AddGrpcService<IVacationRequest>();
+
+        services.AddGrpcService<IAuthenticationService>();
+
+        services.AddIdentityUIServices();
     }
 }
