+++ Avhrm.UI.Shared/appsettings.json
@@ -7,6 +7,6 @@
     }
   },
   "Api": {
-    "Ip": "localhost:5241"
+    "Ip": "http://localhost:5072"
   }
 }
