+++ Avhrm.sln
@@ -11,7 +11,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastru
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UI", "UI", "{FFF5DE75-402B-4245-B562-0C4D173498AE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avhrm.Identity", "Avhrm\Avhrm.Identity.csproj", "{BD075CE9-D02F-461C-8DC8-0FE2313DEBDA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avhrm.Identity.UI", "Avhrm\Avhrm.Identity.UI.csproj", "{BD075CE9-D02F-461C-8DC8-0FE2313DEBDA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avhrm.Persistence", "Avhrm.Persistence\Avhrm.Persistence.csproj", "{D61D6484-0E8E-4AF9-8B98-11B63EE63667}"
 EndProject
@@ -36,7 +36,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avhrm.UI.Source", "Avhrm.UI
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avhrm.UI.Web", "Avhrm.UI.Web\Avhrm.UI.Web.csproj", "{6FA787B2-7D93-4C2C-80E8-DF4D2A205CC7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Avhrm.Tools", "Avhrm.Tools\Avhrm.Tools.csproj", "{DFD9C6D3-A112-4BB9-9349-F8EA3F40A058}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avhrm.Tools", "Avhrm.Tools\Avhrm.Tools.csproj", "{DFD9C6D3-A112-4BB9-9349-F8EA3F40A058}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Avhrm.Identity.Server", "Avhrm.Identity.Server\Avhrm.Identity.Server.csproj", "{A31BE17D-4E5E-4D5A-90D6-8AB5C03F358A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -86,6 +88,10 @@ Global
 		{DFD9C6D3-A112-4BB9-9349-F8EA3F40A058}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DFD9C6D3-A112-4BB9-9349-F8EA3F40A058}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DFD9C6D3-A112-4BB9-9349-F8EA3F40A058}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A31BE17D-4E5E-4D5A-90D6-8AB5C03F358A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A31BE17D-4E5E-4D5A-90D6-8AB5C03F358A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A31BE17D-4E5E-4D5A-90D6-8AB5C03F358A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A31BE17D-4E5E-4D5A-90D6-8AB5C03F358A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -105,6 +111,7 @@ Global
 		{0655AF85-9D05-4008-8B71-D5A6F5AF1C7B} = {FFF5DE75-402B-4245-B562-0C4D173498AE}
 		{6FA787B2-7D93-4C2C-80E8-DF4D2A205CC7} = {FFF5DE75-402B-4245-B562-0C4D173498AE}
 		{DFD9C6D3-A112-4BB9-9349-F8EA3F40A058} = {2257A416-27ED-4E90-8FFE-182B6BF17223}
+		{A31BE17D-4E5E-4D5A-90D6-8AB5C03F358A} = {2257A416-27ED-4E90-8FFE-182B6BF17223}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BFAC11BB-ECB6-4E53-B974-1F2417D73229}
