--- Avhrm/Contracts/IAuthenticationService.cs
@@ -1,7 +1,0 @@
-﻿namespace Avhrm.Identity.Contracts;
-
-public interface IAuthenticationService
-{
-    Task<bool> Authenticate(string username, string password);
-    Task Logout();
-}
